Refactor: move app and config packages to internal for better encapsulation
