+++ cmd/main.go
@@ -1,11 +1,11 @@
 package main
 
 import (
-	"github.com/bohexists/task-manager-svc/app"
-	"github.com/bohexists/task-manager-svc/config"
 	"github.com/bohexists/task-manager-svc/internal/adapters/db"
 	"github.com/bohexists/task-manager-svc/internal/adapters/grpc"
 	"github.com/bohexists/task-manager-svc/internal/adapters/nats"
+	"github.com/bohexists/task-manager-svc/internal/app"
+	"github.com/bohexists/task-manager-svc/internal/config"
 	shutdown "github.com/bohexists/task-manager-svc/internal/system"
 	"github.com/bohexists/task-manager-svc/ports/inbound"
 	"log"
